Restore low quality Cog emblems
In the v2.8.0 update, the TTR team updated the Cog emblems using their HD remasters from years back.

Unfortunately due to an optimization choice made by the team, the emblems that appear inside of Cog building elevators cannot be pixel-perfect to their original counterparts. This is because they have replaced the usual JPG+RGB texture combination with a model, which sadly locks what would've been the RGB file in place, meaning you cannot texture outside of the predefined area anymore.

Luckily for me this doesn't have too much of an impact on bringing the old emblems back since they resemble the same shape, but if you inspect the emblems up-close you may notice imperfections. There's nothing I can do about this, unless the TTR team either revert their decision to swap the emblems to models, or they implement an alternative method for content pack creators that allows them to edit the emblems further again.

NOTE: The model variant of the emblems to the best of my knowledge aren't used anywhere outside of the Cog building elevator, so the rest of emblems should be pixel-perfect. I am unable to dedicate a chunk of time to fully testing these changes however, so if you notice any issues please contact me with evidence.