Optimise teleportation hole texture + RGB
It now uses a much closer resolution to the original TTO teleportation hole RGB, which was 32x32. I made some tweaks to the overall placements to match TTO since copying it 1:1 would result in some odd visual artefacts. The file sizes are also much smaller in comparison, which is a win/win in my book.