Much more accurate PaT background! (Thanks, Ryno!)
Not only is this background much closer to the original Toontown Online Pick-a-Toon background, but it is also smaller in size and works perfectly still. A must-have change for my pack!

Huge thank you to Sarah (RynoGG) for allowing me to use her edit in my content pack! You can find her on Twitter here: https://twitter.com/_rynogg